Remove unnecessary casts
git-svn-id: http://google-refine.googlecode.com/svn/trunk@2173 7d457c2a-affb-35e4-300a-418c747d4874

+++ main/src/com/google/refine/browsing/facets/ScatterplotDrawingRowVisitor.java
@@ -86,7 +86,7 @@ public class ScatterplotDrawingRowVisitor implements RowVisitor, RecordVisitor {
         this.dim_y = dim_y;
         this.rotation = rotation;
         
-        l = (double) size;
+        l = size;
         r = ScatterplotFacet.createRotationMatrix(rotation, l);
 
         image = new BufferedImage(size, size, BufferedImage.TYPE_4BYTE_ABGR);
